requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
branca==0.7.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy @ file:///D:/bld/debugpy_1707444547228/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
distro==1.9.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
flatbuffers==23.5.26
folium==0.16.0
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython @ file:///D:/bld/ipython_1709559926914/work
jax==0.4.25
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///D:/bld/jupyter_core_1704727156030/work
kiwisolver==1.4.5
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.24
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mediapipe==0.10.10
ml-dtypes==0.3.2
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy==1.26.4
openai==1.13.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.15
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.2.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
peewee==3.17.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1705722518504/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq @ file:///D:/bld/pyzmq_1701783318166/work
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
seaborn==0.13.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.31.1
streamlit-folium==0.18.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado @ file:///D:/bld/tornado_1708363274019/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
xyzservices==2023.10.1
yarl==1.9.4
yfinance==0.2.37
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
